fix integer type issue

+++ src/nemos/glm_hmm/expectation_maximization.py
@@ -720,7 +720,9 @@ def max_sum(
         def reset_chain(omega_prev, log_em):
             # New session: reset to initial distribution
             omega = log_init + log_em
-            max_prob_state = jnp.full(n_states, -1, dtype=jnp.int32)  # Boundary marker
+            max_prob_state = jnp.full(
+                n_states, -1, dtype=jnp.integer
+            )  # Boundary marker
             return omega, max_prob_state
 
         def continue_chain(omega_prev, log_em):
@@ -755,7 +757,7 @@ def max_sum(
 
         def session_boundary(state_idx, max_prob, omega):
             # Hit a session start, pick best state at this boundary
-            return jnp.argmax(omega)
+            return jnp.argmax(omega).astype(jnp.integer)
 
         def continue_backward(state_idx, max_prob, omega):
             # Normal backtracking
